+++ frontend/src/App.jsx
@@ -2,13 +2,14 @@ import { useState } from 'react'
 import { BrowserRouter, Routes, Route } from 'react-router'
 import './App.css'
 import Landing from './components/landing/Landing.jsx'
+import Dashboard from './components/Dashboard/Dashboard.jsx'
 
 function App() {
   return (
     <BrowserRouter>
       <Routes>
         <Route path="/" element={<Landing />} />
-        {/*<Route path="/dashboard/" element={<DashboardProfile />} />*/}
+        <Route path="/dashboard/" element={<Dashboard />} />
       </Routes>
     </BrowserRouter>
   )